Fix error preventing SB from building (thanks @norbert!)
