+++ lib/components/src/blocks/EmptyBlock.tsx
@@ -1,7 +1,7 @@
 import React, { FunctionComponent } from 'react';
 import { styled } from '@storybook/theming';
 import { transparentize } from 'polished';
-import { withReset } from '../typography/DocumentFormatting';
+import { withReset } from '../typography/withReset';
 
 const Wrapper = styled.div<{}>(withReset, ({ theme }) => ({
   backgroundColor: theme.base === 'light' ? 'rgba(0,0,0,.01)' : 'rgba(255,255,255,.01)',
