+++ lib/components/src/typography/withReset.tsx
@@ -1,0 +1,12 @@
+import { CSSObject, Theme } from '@storybook/theming';
+
+export const withReset = ({ theme }: { theme: Theme }): CSSObject => ({
+  fontFamily: theme.typography.fonts.base,
+  fontSize: theme.typography.size.s3,
+  margin: 0,
+
+  WebkitFontSmoothing: 'antialiased',
+  MozOsxFontSmoothing: 'grayscale',
+  WebkitTapHighlightColor: 'rgba(0, 0, 0, 0)',
+  WebkitOverflowScrolling: 'touch',
+});
